% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootGSEA.r
\name{aggr.boot.GO}
\alias{aggr.boot.GO}
\title{Aggregate boostrap GO analysis}
\usage{
aggr.boot.GO(resbootGO)
}
\arguments{
\item{resbootGO}{A list as generated by the function boot.GO}
}
\value{
A matrix containing the following columns: GO IDs, rank aggregation score, ranks from the non-bootstrap GO analysis, B colums with ranks from the bootstrap analyses
}
\description{
Aggregation of bootstrap GO analyses
}
\details{
Aggregates ranks from B bootstrap GO analyses
}
\examples{
require(RobustRankAggreg)
require(topGO)
d <- 10000 ### number of genes
sigGenes.example <- rbinom(d, 1, 0.1)
names(sigGenes.example) <- paste("gene", 1:d, sep="_")
gene2GO.example <- vector(mode="list", length=d)
names(gene2GO.example) <- names(sigGenes.example)
GOs = paste("GO:000", 1000:9999, sep="")
for (j in 1:d) gene2GO.example[[j]] <- sample(GOs, rnbinom(1, 1, 0.1),replace=FALSE)
resultGO = boot.GO(sigGenes=sigGenes.example, gene2GO=gene2GO.example)
resultAggr = aggr.boot.GO(resbootGO=resultGO)
}
\references{
Kolde, R., Laur, S., Adler, P., & Vilo, J. (2012). Robust rank aggregation for gene list integration and meta-analysis. \emph{Bioinformatics}, \strong{28(4)}, 573-580. \url{https://doi.org/10.1093/bioinformatics/btr709}
}
\author{
Shamini Hemandhar Kumar, Klaus Jung
}
