% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootGSEA.r
\name{aggr.boot.Pathway}
\alias{aggr.boot.Pathway}
\title{Aggregate boostrap pathway analysis}
\usage{
aggr.boot.Pathway(resbootpathway)
}
\arguments{
\item{resbootpathway}{A list as generated by the function boot.pathway}
}
\value{
A matrix containing the following columns: IDs, rank aggregation score, ranks from the non-bootstrap pathway analysis, B columns with ranks from the bootstrap analyses
}
\description{
Aggregation of bootstrap pathway analyses
}
\details{
Aggregates ranks from B bootstrap pathway analyses
}
\examples{
require(RobustRankAggreg)
require(clusterProfiler)
require(ReactomePA)
genelist <- rnorm(10000) #example for boot.pathway
names(genelist) <- paste( 10000:19999, sep="")
resultpathway = boot.pathway(sigGenes=genelist , B=2, tau=0.95,
 org = 'mouse', pathway = 'Reactome')
resultAggr = aggr.boot.Pathway(resbootpathway = resultpathway)
compareRank(aggrbootGO=resultAggr, lim=100, ord="original", ident=FALSE)
}
\references{
Kolde, R., Laur, S., Adler, P., & Vilo, J. (2012). Robust rank aggregation for gene list integration and meta-analysis. \emph{Bioinformatics}, \strong{28(4)}, 573-580. \url{https://doi.org/10.1093/bioinformatics/btr709}
}
\author{
Shamini Hemandhar Kumar, Klaus Jung
}
