% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootGSEA.r
\name{boot.pathway}
\alias{boot.pathway}
\title{Bootstrap Pathway analysis}
\usage{
boot.pathway(sigGenes, B = 2, tau = 0.95, org = "mouse", pathway = "Reactome")
}
\arguments{
\item{sigGenes}{Vector coding for entrez ids with logfc values}

\item{B}{Number of bootstrap runs}

\item{tau}{Percentage (between 0 and 1) of actual size of genelist to be sampled for bootstrap analysis}

\item{org}{Organism of the genelist to be analysed}

\item{pathway}{One of 'KEGG', 'Reactome' and 'Wiki' pathways}
}
\value{
A list containing the following items:
\describe{
\item{nonbootresult}{Table with results of standard enrichment analysis.}
\item{bootresult}{List of size B, where each element is the result of a bootstrap pathway analysis.}
}
}
\description{
Bootstrap Pathway analysis
}
\details{
Performs bootstrap Pathway analysis for KEGG, Reactome and Wiki Pathways
}
\examples{
require(RobustRankAggreg)
require(clusterProfiler)
require(ReactomePA)
genelist <- rnorm(10000) #example for boot.pathway
names(genelist) <- paste( 10000:19999, sep="")
resultpathway = boot.pathway(sigGenes=genelist , B=2, tau=0.95,
 org = 'mouse', pathway = 'Reactome')
resultAggr = aggr.boot.Pathway(resbootpathway = resultpathway)
compareRank(aggrbootGO=resultAggr, lim=100, ord="original", ident=FALSE)
}
\references{
T Wu, E Hu, S Xu, M Chen, P Guo, Z Dai, T Feng, L Zhou, W Tang, L Zhan, X Fu, S Liu, X Bo, and G Yu. clusterProfiler 4.0: A universal enrichment tool for interpreting omics data. The Innovation. 2021, 2(3):100141 \url{https://bioconductor.org/packages/release/bioc/html/clusterProfiler.html}
Guangchuang Yu, Qing-Yu He. ReactomePA: an R/Bioconductor package for reactome pathway analysis and visualization.Molecular BioSystems 2016, 12(2):477-479 \url{https://bioconductor.org/packages/release/bioc/html/ReactomePA.html}
}
\author{
Shamini Hemandhar Kumar, Klaus Jung
}
