% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootGSEA.r
\docType{package}
\name{bootGSEA}
\alias{bootGSEA}
\title{Robustness evaluation of gene set enrichment analysis (GSEA)}
\description{
Package contains functions that repeates GSEA using bootstrap samples of gene sets. Bootstrap results are
aggregated to a new ranking score. The score can be compared to the gene set
ranking resulting from the standard GSEA.
}
\details{
So far the functions included in the package are:
\itemize{
\item \code{\link{aggr.boot.GO}} Aggregate boostrap GO analysis
\item \code{\link{aggr.boot.Pathway}} Write a \code{genind} Aggregate boostrap pathway analysis
\item \code{\link{aggr.multiomics}} Multiomics Integration analysis
\item \code{\link{boot.GO}} Bootstrap GO analysis
\item \code{\link{boot.pathway}} Bootstrap Pathway analysis
\item \code{\link{compareRank}} Visualisation of bootstrap GO analyses
\item \code{\link{histDiff}} Visualization of difference in ranks
\item \code{\link{plotRank}} Visualisation of bootstrap pathway analyses
}

\tabular{ll}{
Package: \tab bootGSEA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2023-04-04\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Shamini Hemandhar Kumar, Klaus Jung (\email{shamini.hemandhar.kumar@tiho-hannover.de})(\email{klaus.jung@tiho-hannover.de})

Maintainer: Shamini Hemandhar Kumar (\email{shamini.hemandhar.kumar@tiho-hannover.de})
}
\keyword{GSEA}
\keyword{Robustness}
