% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootGSEA.r
\name{compareRank}
\alias{compareRank}
\title{Visualisation of bootstrap GO analyses}
\usage{
compareRank(aggrbootGO, lim = 100, ord = "original", ident = FALSE)
}
\arguments{
\item{aggrbootGO}{A table as returned by the function aggr.boot.GO}

\item{lim}{Number of top ranked GO terms to be displayed in the plot}

\item{ord}{Character with value "original" for ranking GO terms according to non-bootstrap GO analysis}

\item{ident}{Boolean value. If true, the user is enabled to click and annotate GO terms in the plot}
}
\value{
No return value
}
\description{
Visualisation of bootstrap GO analyses
}
\details{
Generates a plot for visualisation of bootstrap GO analyses
}
\examples{
require(RobustRankAggreg)
require(topGO)
d <- 10000 ### number of genes
sigGenes.example <- rbinom(d, 1, 0.1)
names(sigGenes.example) <- paste("gene", 1:d, sep="_")
gene2GO.example <- vector(mode="list", length=d)
names(gene2GO.example) <- names(sigGenes.example)
GOs = paste("GO:000", 1000:9999, sep="")
for (j in 1:d) gene2GO.example[[j]] <- sample(GOs, rnbinom(1, 1, 0.1),replace=FALSE)
resultGO = boot.GO(sigGenes=sigGenes.example, gene2GO=gene2GO.example)
resultAggr = aggr.boot.GO(resbootGO=resultGO)
compareRank(aggrbootGO=resultAggr, lim=100, ord="original", ident=FALSE)
}
\references{
Alexa, A., & Rahnenführer, J. (2009). Gene set enrichment analysis with topGO. \emph{Bioconductor Improv}, strong{27}, 1-26. \url{https://mirrors.nju.edu.cn/bioconductor/3.2/bioc/vignettes/topGO/inst/doc/topGO.pdf}
Kolde, R., Laur, S., Adler, P., & Vilo, J. (2012). Robust rank aggregation for gene list integration and meta-analysis. \emph{Bioinformatics}, \strong{28(4)}, 573-580. \url{https://doi.org/10.1093/bioinformatics/btr709}
}
\author{
Shamini Hemandhar Kumar, Klaus Jung
}
