% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootGSEA.r
\name{histDiff}
\alias{histDiff}
\title{Visualization of difference in ranks}
\usage{
histDiff(
  standardRank,
  bootstrappedRank,
  title = "",
  xlab = "",
  ylab = "",
  binwidth = 5
)
}
\arguments{
\item{standardRank}{The standard rank column of omics analysis}

\item{bootstrappedRank}{The bootstrapped rank column of omics analysis}

\item{title}{Title of the histogram}

\item{xlab}{Title of the xaxis}

\item{ylab}{Title of the yaxis}

\item{binwidth}{Width of the bins}
}
\value{
No return value
}
\description{
Visualization of rank difference
}
\details{
Visualizes difference in ranks between standard and bootstrapped analyses
}
\examples{
require(RobustRankAggreg)
require(topGO)
require(ggplot2)
require(dplyr)
d <- 10000 ### number of genes
sigGenes.example <- rbinom(d, 1, 0.1)
names(sigGenes.example) <- paste("gene", 1:d, sep="_")
gene2GO.example <- vector(mode="list", length=d)
names(gene2GO.example) <- names(sigGenes.example)
GOs = paste("GO:000", 1000:9999, sep="")
for (j in 1:d) gene2GO.example[[j]] <- sample(GOs, rnbinom(1, 1, 0.1),replace=FALSE)
resultGO = boot.GO(sigGenes=sigGenes.example, gene2GO=gene2GO.example)
resultAggr = aggr.boot.GO(resbootGO=resultGO)
stdrank <- resultAggr$rankRun_nonBoot
bootrank <- rank(resultAggr$Score)
histDiff(standardRank = stdrank, bootstrappedRank = bootrank, title = "",
xlab = "Rank difference (standard-bootstrap)", ylab = "Frequency", binwidth = 15)
}
\references{
Wickham H (2016). \emph{ggplot2: Elegant Graphics for Data Analysis}. Springer-Verlag New York. ISBN 978-3-319-24277-4, \url{https://ggplot2.tidyverse.org}
Wickham H, François R, Henry L, Müller K (2022). \emph{dplyr: A Grammar of Data Manipulation}. R package version 1.0.9, \url{https://CRAN.R-project.org/package=dplyr}
}
\author{
Shamini Hemandhar Kumar, Klaus Jung
}
