% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootGSEA.r
\name{plotRank}
\alias{plotRank}
\title{Visualisation of bootstrap pathway analyses}
\usage{
plotRank(
  aggrboot,
  ordercolumn,
  ColumnA,
  ColumnB,
  ord = "standard",
  title = "",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{aggrboot}{A table with values for plot}

\item{ordercolumn}{The column with which the table should be ordered}

\item{ColumnA}{The rank column which has to be plotted along the x-axis}

\item{ColumnB}{The rank column which has to be plotted along the y-axis}

\item{ord}{Character with value "standard" for ranking pathways according to ordercolumn}

\item{title}{The title of the plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
No return value
}
\description{
Visualisation of bootstrap analyses
}
\details{
Generates a plot for visualisation of bootstrap pathway analyses
}
\examples{
require(RobustRankAggreg)
require(topGO)
require(dplyr)
require(clusterProfiler)
require(ReactomePA)
require(ggplot2)
genelist_T <- rnorm(10000)
names(genelist_T) <- paste( 10000:19999, sep="")
genelist_P <- rnorm(10000)
names(genelist_P) <- paste( 10000:19999, sep="")
resultpathway_T = boot.pathway(sigGenes=genelist_T , B=2, tau=0.95,
 org = 'mouse', pathway = 'Reactome')
resultAggr_T = aggr.boot.Pathway(resbootpathway = resultpathway_T)
resultpathway_P = boot.pathway(sigGenes=genelist_P , B=2, tau=0.95,
 org = 'mouse', pathway = 'Reactome')
resultAggr_P = aggr.boot.Pathway(resbootpathway = resultpathway_P)
Compareomics = aggr.multiomics(resultAggr_T = resultAggr_T,
 resultAggr_P = resultAggr_P)
Compareomics <- left_join(Compareomics,
resultAggr_T \%>\% dplyr::select(Name,rankRun_nonBoot, Score),
by = "Name")
colnames(Compareomics) = c("Name", "IntegratedScore",
 "RankTrans", "RankProt", "rankRun_nonBoot_T", "Score_T")
Compareomics <- left_join(Compareomics,
resultAggr_P \%>\% dplyr::select(Name,rankRun_nonBoot, Score),by = "Name")
colnames(Compareomics) = c("Name", "IntegratedScore", "RankTrans", "RankProt",
 "rankRun_nonBoot_T", "Score_T", "rankRun_nonBoot_P", "Score_P")
plotRank(aggrboot = Compareomics, ordercolumn = "IntegratedScore",
 ColumnA = "rankRun_nonBoot_T", ColumnB = "RankTrans", title = "Rank comparison",
  xlab = "Standard Enrichment analysis", ylab = "Bootstrapped enrichment analysis")
}
\references{
Wickham H (2016). \emph{ggplot2: Elegant Graphics for Data Analysis}. Springer-Verlag New York. ISBN 978-3-319-24277-4, \url{https://ggplot2.tidyverse.org}
Wickham H, François R, Henry L, Müller K (2022). \emph{dplyr: A Grammar of Data Manipulation}. R package version 1.0.9, \url{https://CRAN.R-project.org/package=dplyr}
}
\author{
Shamini Hemandhar Kumar, Klaus Jung
}
